requirements.txt
alembic==1.8.1
anyio==3.6.1
bcrypt==4.0.0
black==22.8.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.85.0
flake8==5.0.4
greenlet==1.1.3
h11==0.13.0
idna==3.4
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycodestyle==2.9.1
pydantic==1.10.2
pyflakes==2.5.0
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
